+++ project/src/main/java/com/favorite/project/controller/ClosetController.java
@@ -1,4 +1,4 @@
-package com.favorite.project.controller.api;
+package com.favorite.project.controller;
 
 import com.favorite.project.dao.ClosetDao;
 import com.favorite.project.entity.Closet;
