+++ project/src/main/java/com/favorite/project/controller/ClosetPageController.java
@@ -1,4 +1,4 @@
-package com.favorite.project.controller.view;
+package com.favorite.project.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
