+++ project/src/main/resources/templates/getClosetList.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>여기 옷장들 모임이야</title>
+</head>
+<body>
+
+
+</body>
+</html>
